+++ CHANGELOG.md
@@ -5,11 +5,9 @@ All Notable changes to `Slytherin` will be documented in this file
 ## [Unreleased](https://github.com/rougin/slytherin/compare/v0.7.0...HEAD) - [CURRENT_DATE]
 
 ### Added
+- `BaseUriGuesser` for filtering the base URI
 - Unit test for `Phroute` in `Application\Application`
 - Middlewares in `FastRoute\Dispatcher` and ``Phroute\Dispatcher`
-
-### Fixed
-- Route error when accessing application in localhost (`127.0.0.1`, `::1`)
 
 ## [0.8.0](https://github.com/rougin/slytherin/compare/v0.7.0...v0.8.0) - 2016-09-08